pluribus/requirements_compatible.txt
blessed>=1.17.9
click>=7.0
dill>=0.3.1.1
enlighten>=1.5.2
ipdb>=0.12.3
matplotlib>=3.7.0
numpy>=1.24.0
pytest>=5.4.3
PyYAML>=5.3.1
rich>=1.0.0
recommonmark>=0.6.0
scikit-learn>=1.3.0
scipy>=1.10.0
Sphinx>=3.0.4
sphinx-rtd-theme>=0.4.3
tqdm>=4.43.0
memory_profiler>=0.61.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
joblib>=1.0.0 